chore:  renamed StaticLegend to StaticLegendBox for consistency.  no functional changes
