+++ giraffe/src/components/StaticLegendBox.tsx
